Seperate out the interaction model. Move the lambda function
This fixes all the response logic being baked into the lambda_func
module. This was so overtime, issues with the dialog can be addressed as
they are reported by users. There are still some dialog issues, but now
the code that controls the alexa responses is in the interaction model,
somewhat seperate from the code that handles the front-end to lambda. In
addition, there is one env module now that is mostly used to deal with
time correctly...

I need to write a lot of unit tests but I'm kind of skipping that for
now..