Log and raise custom exception when an unknown intent comes in
Because there are some cloudwatch alarms setup for lambda function
failures, this might make it easier to figure out what happened if the
function crashes. I was going to return a message to the user, but I
think that I really don't have a good answer for what to do right here
and crashing and notifying me ASAP is the right course of action.